+++ frontend/src/components/SignIn/index.tsx
@@ -1,20 +1,68 @@
 import { useState } from 'react';
-import Modal from '../Modal/Modal';
+import isUserExist from '../../utils/isUserExist';
+import Input from '../Input';
+import isCorrectPswd from '../../utils/isCorrectPswd';
+import setSession from '../../utils/setSession';
+import {
+  closeModal,
+  fetchFavorites, fetchHistory, setUser, useAppDispatch,
+} from '../../store';
 import './SignIn.scss';
-import SignInForm from './SignInForm';
-import { openModal, useAppDispatch } from '../../store';
 
 function SignIn() {
   const dispatch = useAppDispatch();
-  const [active, setActive] = useState(false);
-  const handleClick = () => dispatch(openModal('signin'));
+  const [loginErr, setLoginErr] = useState('');
+  const [loginValue, setLoginValue] = useState('');
+  const [passErr, setPassErr] = useState('');
+  const [passValue, setPassValue] = useState('');
+  const handleSubmit = async () => {
+    if (!isUserExist(loginValue)) {
+      setPassValue('');
+      setLoginErr('Account not found');
+      return null;
+    }
+    const correctPswd = await isCorrectPswd(loginValue, passValue);
+    if (!correctPswd) {
+      setPassErr('incorrect password');
+      setPassValue('');
+      return null;
+    }
+    await setSession(loginValue);
+    dispatch(setUser());
+    dispatch(fetchHistory());
+    dispatch(fetchFavorites());
+    dispatch(closeModal());
+    return null;
+  };
+
   return (
-    <>
-      <button type="button" className="signin__button" onClick={handleClick}>SignIn</button>
-      <Modal setActive={setActive} active={active}>
-        <SignInForm setActive={setActive} />
-      </Modal>
-    </>
+    <div className="signin__form">
+      <div className="signin__title">
+        authorization
+      </div>
+      <Input
+        error={loginErr}
+        value={loginValue}
+        placeholder="login"
+        onChange={({ target }) => setLoginValue(target.value.trim())}
+        onFocus={() => setLoginErr('')}
+      />
+      <Input
+        password
+        error={passErr}
+        value={passValue}
+        placeholder="password"
+        onChange={({ target }) => setPassValue(target.value.trim())}
+        onFocus={() => setPassErr('')}
+      />
+      <button
+        className="signin__formbutton_submit"
+        onClick={handleSubmit}
+        type="button"
+      >
+        SignIn
+      </button>
+    </div>
   );
 }
